[colorwheel] add verification tolerance and address potential image size overflow

+++ src/colorwheel-sycl/main.cpp
@@ -71,7 +71,7 @@ void color (sycl::nd_item<2> &item, uchar* pix, int size, int half_size, float r
     float fx = (float)x / (float)half_size * range - range;
     float fy = (float)y / (float)half_size * range - range;
     if (x == half_size || y == half_size) return; // make black coordinate axes
-    size_t idx = (y * size + x) * 3;
+    size_t idx = ((size_t)y * size + x) * 3;
     computeColor(fx/truerange, fy/truerange, pix+idx);
   }
 }
@@ -92,7 +92,7 @@ int main(int argc, char **argv)
   const int half_size = size/2;
 
   // create a test image showing the color encoding
-  size_t imgSize = size * size * 3;
+  size_t imgSize = (size_t)size * size * 3;
   uchar* pix = (uchar*) malloc (imgSize);
   uchar* res = (uchar*) malloc (imgSize);
 
@@ -103,7 +103,7 @@ int main(int argc, char **argv)
       float fx = (float)x / (float)half_size * range - range;
       float fy = (float)y / (float)half_size * range - range;
       if (x == half_size || y == half_size) continue; // make black coordinate axes
-      size_t idx = (y * size + x) * 3;
+      size_t idx = ((size_t)y * size + x) * 3;
       computeColor(fx/truerange, fy/truerange, pix+idx);
     }
   }
